Clean up the confidence interval framework. (#361)
* Store ci_group_size as part of the forest to avoid a mismatch between training and prediction.
* Add an estimate_variance option to prediction, to replace the fragile check ci_group_size > 1.
* Move the check for consistent ci_group_size during merges into C++.
* Use size_t instead of uint for ci_group_size.

+++ core/src/prediction/LocalLinearPredictionStrategy.h
@@ -54,7 +54,7 @@ public:
         std::vector<std::vector<size_t>> samples_by_tree,
         std::unordered_map<size_t, double> weights_by_sampleID,
         const Observations& observations,
-        uint ci_group_size);
+        size_t ci_group_size);
 
 private:
